Working draft of key features, including emojistats and cache.
This includes a working first draft of several key voobot features.

In particular, it includes a simple interface for counting, tabulating,
and displaying the number of occurrences of each particular emote that
may appear in a reaction.

This commit also includes a TinyDB cache of Discord models to
minimize the number of requests that must go through the Discord API.

This commit melds together several earlier, messier, prototyping
commits into a single cleaner-looking package.